requirements.txt
selenium
pytest
pytest-html
pyyaml
webdriver-manager